+++ tools/circt-reduce/CMakeLists.txt
@@ -4,6 +4,7 @@ set(LLVM_LINK_COMPONENTS
 
 add_llvm_tool(circt-reduce
  circt-reduce.cpp
+ Reduction.cpp
 )
 llvm_update_compile_flags(circt-reduce)
 target_link_libraries(circt-reduce
